requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.28.0
alembic==1.12.1
pydantic==2.5.2
pydantic-settings==2.0.3
paho-mqtt==1.6.1
redis==5.0.1
python-dotenv==1.0.0
greenlet==3.0.1